+++ integration/vite-spa-mode-test.ts
@@ -637,6 +637,61 @@ test.describe("SPA Mode", () => {
           expect(html.match(/window.__reactRouterContext =/g)?.length).toBe(1);
           expect(html.match(/💿 Hey developer 👋/g)?.length).toBe(1);
         });
+
+        test("does not inherit single fetch revalidation behavior", async ({
+          page,
+        }) => {
+          fixture = await createFixture({
+            spaMode: true,
+            files: {
+              "react-router.config.ts": reactRouterConfig({
+                ssr: false,
+                splitRouteModules,
+              }),
+              "app/routes/_index.tsx": js`
+                import { Link } from 'react-router';
+                export default function Component() {
+                  return <Link to="/parent">Go to parent</Link>;
+                }
+              `,
+              "app/routes/parent.tsx": js`
+                import { Link, Outlet } from 'react-router';
+                let count = 0;
+                export function clientLoader() {
+                  return ++count;
+                }
+                export default function Component({ loaderData }) {
+                  return (
+                    <>
+                      <h1>Parent: {loaderData}</h1>
+                      <Link to="./child">Go to child</Link>
+                      <Outlet />
+                    </>
+                  )
+                }
+              `,
+              "app/routes/parent.child.tsx": js`
+                import { Link } from 'react-router';
+                export default function Component({ loaderData }) {
+                  return <h2>Child</h2>;
+                }
+              `,
+            },
+          });
+          appFixture = await createAppFixture(fixture);
+
+          let app = new PlaywrightFixture(appFixture, page);
+          await app.goto("/", true);
+          await page.waitForSelector('a[href="/parent"]');
+          await app.clickLink("/parent");
+          await page.waitForSelector("h1");
+          expect(await page.locator("h1").textContent()).toBe("Parent: 1");
+
+          await app.clickLink("/parent/child");
+          await page.waitForSelector("h2");
+          expect(await page.locator("h1").textContent()).toBe("Parent: 1");
+          expect(await page.locator("h2").textContent()).toBe("Child");
+        });
       });
 
       test.describe("normal apps", () => {
@@ -850,7 +905,7 @@ test.describe("SPA Mode", () => {
           appFixture.close();
         });
 
-        test("renders the root HydrateFallback initially with access to the root loader data", async ({}) => {
+        test("renders the root HydrateFallback initially with access to the root loader data", async () => {
           let res = await fixture.requestDocument("/");
           let html = await res.text();
           expect(html).toMatch('<h1 data-loading="true">Loading SPA...</h1>');
@@ -984,7 +1039,7 @@ test.describe("SPA Mode", () => {
           await app.clickLink("/error");
           await page.waitForSelector("[data-error]");
           expect(await page.locator("[data-error]").textContent()).toBe(
-            'Error: You cannot call serverLoader() in SPA Mode (routeId: "routes/error")'
+            'Error: You are trying to call serverLoader() on a route that does not have a server loader (routeId: "routes/error")'
           );
         });
 
@@ -998,7 +1053,7 @@ test.describe("SPA Mode", () => {
           await app.clickSubmitButton("/error");
           await page.waitForSelector("[data-error]");
           expect(await page.locator("[data-error]").textContent()).toBe(
-            'Error: You cannot call serverAction() in SPA Mode (routeId: "routes/error")'
+            'Error: You are trying to call serverAction() on a route that does not have a server action (routeId: "routes/error")'
           );
         });
 
@@ -1011,5 +1066,116 @@ test.describe("SPA Mode", () => {
         });
       });
     });
+  });
+
+  test("only imports the root route in the server build when SSRing index.html", async ({
+    page,
+  }) => {
+    let fixture = await createFixture({
+      spaMode: true,
+      files: {
+        "react-router.config.ts": reactRouterConfig({
+          ssr: false,
+        }),
+        "vite.config.ts": js`
+          import { defineConfig } from "vite";
+          import { reactRouter } from "@react-router/dev/vite";
+
+          export default defineConfig({
+            build: { manifest: true },
+            plugins: [reactRouter()],
+          });
+        `,
+        "app/routeImportTracker.ts": js`
+          // this is kinda silly, but this way we can track imports
+          // that happen during SSR and during CSR
+          export async function logImport(url: string) {
+            try {
+              const fs = await import("node:fs");
+              const path = await import("node:path");
+              fs.appendFileSync(path.join(process.cwd(), "ssr-route-imports.txt"), url + "\n");
+            }
+            catch (e) {
+              (window.csrRouteImports ??= []).push(url);
+            }
+          }
+        `,
+        "app/root.tsx": js`
+          import { Links, Meta, Outlet, Scripts } from "react-router";
+          import { logImport } from "./routeImportTracker";
+          logImport("app/root.tsx");
+
+          export default function Root() {
+            return (
+              <html lang="en">
+                <head>
+                  <Meta />
+                  <Links />
+                </head>
+                <body>
+                  hello world
+                  <Outlet />
+                  <Scripts />
+                </body>
+              </html>
+            );
+          }
+        `,
+        "app/routes/_index.tsx": js`
+          import { logImport } from "../routeImportTracker";
+          logImport("app/routes/_index.tsx");
+
+          // This should not cause an error on SSR because the module is not loaded
+          console.log(window);
+
+          export default function Component() {
+            return "index";
+          }
+        `,
+        "app/routes/about.tsx": js`
+          import * as React  from "react";
+          import { logImport } from "../routeImportTracker";
+          logImport("app/routes/about.tsx");
+
+          // This should not cause an error on SSR because the module is not loaded
+          console.log(window);
+
+          export default function Component() {
+            const [mounted, setMounted] = React.useState(false);
+            React.useEffect(() => setMounted(true), []);
+
+            return (
+              <>
+                {!mounted ? <span>Unmounted</span> : <span data-mounted>Mounted</span>}
+              </>
+            );
+          }
+        `,
+      },
+    });
+
+    let importedRoutes = (
+      await fs.promises.readFile(
+        path.join(fixture.projectDir, "ssr-route-imports.txt"),
+        "utf-8"
+      )
+    )
+      .trim()
+      .split("\n");
+    expect(importedRoutes).toStrictEqual([
+      "app/root.tsx",
+      // we should not have imported app/routes/_index.tsx
+      // we should not have imported app/routes/about.tsx
+    ]);
+
+    appFixture = await createAppFixture(fixture);
+    let app = new PlaywrightFixture(appFixture, page);
+    await app.goto("/about");
+    await page.waitForSelector("[data-mounted]");
+    // @ts-expect-error
+    expect(await page.evaluate(() => window.csrRouteImports)).toStrictEqual([
+      "app/root.tsx",
+      "app/routes/about.tsx",
+    ]);
   });
 });