+++ packages/react-router/lib/dom/ssr/routeModules.ts
@@ -5,6 +5,7 @@ import type {
   ActionFunctionArgs,
   LoaderFunction,
   LoaderFunctionArgs,
+  unstable_MiddlewareFunction,
   Params,
   ShouldRevalidateFunction,
 } from "../../router/utils";
@@ -24,6 +25,7 @@ export interface RouteModules {
 export interface RouteModule {
   clientAction?: ClientActionFunction;
   clientLoader?: ClientLoaderFunction;
+  unstable_clientMiddleware?: unstable_MiddlewareFunction<undefined>[];
   ErrorBoundary?: ErrorBoundaryComponent;
   HydrateFallback?: HydrateFallbackComponent;
   Layout?: LayoutComponent;
@@ -41,6 +43,7 @@ export interface ServerRouteModule extends RouteModule {
   action?: ActionFunction;
   headers?: HeadersFunction | { [name: string]: string };
   loader?: LoaderFunction;
+  unstable_middleware?: unstable_MiddlewareFunction<Response>[];
 }
 
 /**
@@ -53,7 +56,7 @@ export type ClientActionFunction = (
 /**
  * Arguments passed to a route `clientAction` function
  */
-export type ClientActionFunctionArgs = ActionFunctionArgs<undefined> & {
+export type ClientActionFunctionArgs = ActionFunctionArgs & {
   serverAction: <T = unknown>() => Promise<SerializeFrom<T>>;
 };
 
@@ -69,7 +72,7 @@ export type ClientLoaderFunction = ((
 /**
  * Arguments passed to a route `clientLoader` function
  */
-export type ClientLoaderFunctionArgs = LoaderFunctionArgs<undefined> & {
+export type ClientLoaderFunctionArgs = LoaderFunctionArgs & {
   serverLoader: <T = unknown>() => Promise<SerializeFrom<T>>;
 };